requirements.txt
automata-lib
numpy
graphviz
termcolor
sklearn
tqdm